requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
websockets
python-dotenv
transformers

